LayoutTests/platform/blackberry/accessibility/ellipsis-text-expected.txt

text
My Writing Nook for iPad. More text, more text, more text.
My Writing Nook for iPad. More text, more text, more text.
text
This test makes sure that the bounds of text that overflows with ellipsis is correct (shortened to the ellipsis that is).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL clippedWidth < fullWidth should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
